+++ standalone-scripts/xpath/src/instruction.ts
@@ -11,7 +11,7 @@ import type { ProjectInstruction } from "../../types/instruction/project-instruc
 import type { EmptySettings } from "../../types/instruction/seed/empty-settings";
 
 const instruction: ProjectInstruction<EmptySettings> = {
-    SchemaVersion: "2.240.0",
+    SchemaVersion: "2.241.0",
     Name: "xpath",
     DisplayName: "XPath Utilities",
     Version: "2.239.0",